🚀 Update File Handling in XversE <3...
